+++ src/aks-preview/azext_aks_preview/managed_cluster_decorator.py
@@ -9,6 +9,7 @@ import os
 import semver
 from types import SimpleNamespace
 from typing import Any, Dict, List, Optional, Tuple, TypeVar, Union
+from knack.util import CLIError
 
 from azure.mgmt.containerservice.models import KubernetesSupportPlan
 
@@ -69,6 +70,9 @@ from azext_aks_preview._consts import (
     CONST_NETWORK_DATAPLANE_CILIUM,
     CONST_PRIVATE_DNS_ZONE_NONE,
     CONST_PRIVATE_DNS_ZONE_SYSTEM,
+    CONST_AZURE_KEYVAULT_SECRETS_PROVIDER_ADDON_NAME,
+    CONST_SECRET_ROTATION_ENABLED,
+    CONST_ROTATION_POLL_INTERVAL,
     CONST_AZURE_SERVICE_MESH_UPGRADE_COMMAND_START,
     CONST_AZURE_SERVICE_MESH_UPGRADE_COMMAND_COMPLETE,
     CONST_AZURE_SERVICE_MESH_UPGRADE_COMMAND_ROLLBACK,
@@ -93,7 +97,7 @@ from azext_aks_preview.agentpool_decorator import (
     AKSPreviewAgentPoolUpdateDecorator,
 )
 from azext_aks_preview._roleassignments import add_role_assignment
-from msrestazure.tools import is_valid_resource_id
+from msrestazure.tools import is_valid_resource_id, parse_resource_id
 
 from dateutil.parser import parse
 
@@ -498,7 +502,9 @@ class AKSPreviewManagedClusterContext(AKSManagedClusterContext):
         # read the original value passed by the command
         network_plugin = self.raw_param.get("network_plugin")
         # try to read the property value corresponding to the parameter from the `mc` object
+        # but do not override if it was specified in the raw params since this property can be updated.
         if (
+            not network_plugin and
             self.mc and
             self.mc.network_profile and
             self.mc.network_profile.network_plugin is not None
@@ -607,7 +613,19 @@ class AKSPreviewManagedClusterContext(AKSManagedClusterContext):
 
         :return: bool or None
         """
-        return self.raw_param.get("enable_network_observability")
+        enable_network_observability = self.raw_param.get("enable_network_observability")
+        disable_network_observability = self.raw_param.get("disable_network_observability")
+        if enable_network_observability and disable_network_observability:
+            raise MutuallyExclusiveArgumentError(
+                "Cannot specify --enable-network-observability and "
+                "--disable-network-observability at the same time."
+            )
+        if enable_network_observability is False and disable_network_observability is False:
+            return None
+        if enable_network_observability is not None:
+            return enable_network_observability
+        if disable_network_observability is not None:
+            return not disable_network_observability
 
     def get_load_balancer_managed_outbound_ip_count(self) -> Union[int, None]:
         """Obtain the value of load_balancer_managed_outbound_ip_count.
@@ -1046,6 +1064,16 @@ class AKSPreviewManagedClusterContext(AKSManagedClusterContext):
                 )
 
         return profile
+
+    def get_enable_image_integrity(self) -> bool:
+        """Obtain the value of enable_image_integrity.
+
+        :return: bool
+        """
+        # read the original value passed by the command
+        enable_image_integrity = self.raw_param.get("enable_image_integrity")
+
+        return enable_image_integrity
 
     def get_disable_image_integrity(self) -> bool:
         """Obtain the value of disable_image_integrity.
@@ -1673,10 +1701,10 @@ class AKSPreviewManagedClusterContext(AKSManagedClusterContext):
         """
         return self._get_api_server_authorized_ip_ranges(enable_validation=True)
 
-    def get_dns_zone_resource_ids(self) -> Union[str, None]:
+    def get_dns_zone_resource_ids(self) -> Union[list, None]:
         """Obtain the value of dns_zone_resource_ids.
 
-        :return: string or None
+        :return: list or None
         """
         # read the original value passed by the command
         dns_zone_resource_ids = self.raw_param.get("dns_zone_resource_ids")
@@ -2346,14 +2374,12 @@ class AKSPreviewManagedClusterContext(AKSManagedClusterContext):
 
         :return: string or None
         """
-        # default to None
-        support_plan = None
-        # try to read the property value corresponding to the parameter from the `mc` object
-        if self.mc and hasattr(self.mc, "support_plan") and self.mc.support_plan is not None:
-            support_plan = self.mc.support_plan
-
-        # if specified by customer, use the specified value
+        # take input
         support_plan = self.raw_param.get("k8s_support_plan")
+        if support_plan is None:
+            # user didn't update this property, load from existing ManagedCluster
+            if self.mc and hasattr(self.mc, "support_plan") and self.mc.support_plan is not None:
+                support_plan = self.mc.support_plan
 
         # this parameter does not need dynamic completion
         # this parameter does not need validation
@@ -2400,6 +2426,77 @@ class AKSPreviewManagedClusterContext(AKSManagedClusterContext):
         # Note: No need to check for mutually exclusive parameter with enable-cost-analysis here
         # because it's already checked in _get_enable_cost_analysis
         return self.raw_param.get("disable_cost_analysis")
+
+    def get_keyvault_id(self) -> str:
+        """Obtain the value of keyvault_id.
+
+        :return: str
+        """
+        return self.raw_param.get("keyvault_id")
+
+    def get_enable_kv(self) -> bool:
+        """Obtain the value of enable_kv.
+
+        :return: bool
+        """
+        return self.raw_param.get("enable_kv")
+
+    def get_attach_zones(self) -> bool:
+        """Obtain the value of attach_zones.
+
+        :return: bool
+        """
+        return self.raw_param.get("attach_zones")
+
+    def get_enable_app_routing(self) -> bool:
+        """Obtain the value of enable_app_routing.
+
+        :return: bool
+        """
+        return self.raw_param.get("enable_app_routing")
+
+    def get_dns_zone_resource_ids_from_input(self) -> Union[List[str], None]:
+        """Obtain the value of dns_zone_resource_ids.
+
+        :return: list of str or None
+        """
+        dns_zone_resource_ids = self.raw_param.get("dns_zone_resource_ids")
+        dns_zone_resource_ids = [
+            x.strip()
+            for x in (
+                dns_zone_resource_ids.split(",")
+                if dns_zone_resource_ids
+                else []
+            )
+        ]
+
+        return dns_zone_resource_ids
+
+    def get_add_dns_zone(self) -> bool:
+        """Obtain the value of add_dns_zone.
+
+        :return: bool
+        """
+        return self.raw_param.get("add_dns_zone")
+
+    def get_delete_dns_zone(self) -> bool:
+        """Obtain the value of delete_dns_zone.
+
+        :return: bool
+        """
+        return self.raw_param.get("delete_dns_zone")
+
+    def get_update_dns_zone(self) -> bool:
+        """Obtain the value of update_dns_zone.
+
+        :return: bool
+        """
+        return self.raw_param.get("update_dns_zone")
+
+    def get_node_provisioning_mode(self) -> Union[str, None]:
+        """Obtain the value of node_provisioning_mode.
+        """
+        return self.raw_param.get("node_provisioning_mode")
 
 
 class AKSPreviewManagedClusterCreateDecorator(AKSManagedClusterCreateDecorator):
@@ -2598,6 +2695,22 @@ class AKSPreviewManagedClusterCreateDecorator(AKSManagedClusterCreateDecorator):
             if mc.security_profile is None:
                 mc.security_profile = self.models.ManagedClusterSecurityProfile()
             mc.security_profile.workload_identity = profile
+
+        return mc
+
+    def set_up_image_integrity(self, mc: ManagedCluster) -> ManagedCluster:
+        """Set up security profile imageIntegrity for the ManagedCluster object.
+
+        :return: the ManagedCluster object
+        """
+        self._ensure_mc(mc)
+
+        if self.context.get_enable_image_integrity():
+            if mc.security_profile is None:
+                mc.security_profile = self.models.ManagedClusterSecurityProfile()
+            mc.security_profile.image_integrity = self.models.ManagedClusterSecurityProfileImageIntegrity(
+                enabled=True,
+            )
 
         return mc
 
@@ -2910,6 +3023,39 @@ class AKSPreviewManagedClusterCreateDecorator(AKSManagedClusterCreateDecorator):
 
         return mc
 
+    def set_up_app_routing_profile(self, mc: ManagedCluster) -> ManagedCluster:
+        """Set up app routing profile for the ManagedCluster object.
+
+        :return: the ManagedCluster object
+        """
+        self._ensure_mc(mc)
+
+        if self.context.get_enable_app_routing():
+            if mc.ingress_profile is None:
+                mc.ingress_profile = self.models.ManagedClusterIngressProfile()
+            mc.ingress_profile.web_app_routing = self.models.ManagedClusterIngressProfileWebAppRouting(enabled=True)
+        return mc
+
+    def set_up_node_provisioning_mode(self, mc: ManagedCluster) -> ManagedCluster:
+        self._ensure_mc(mc)
+
+        mode = self.context.get_node_provisioning_mode()
+        if mode is not None:
+            if mc.node_provisioning_profile is None:
+                mc.node_provisioning_profile = self.models.ManagedClusterNodeProvisioningProfile()
+
+            # set mode
+            mc.node_provisioning_profile.mode = mode
+
+        return mc
+
+    def set_up_node_provisioning_profile(self, mc: ManagedCluster) -> ManagedCluster:
+        self._ensure_mc(mc)
+
+        mc = self.set_up_node_provisioning_mode(mc)
+
+        return mc
+
     def construct_mc_profile_preview(self, bypass_restore_defaults: bool = False) -> ManagedCluster:
         """The overall controller used to construct the default ManagedCluster profile.
 
@@ -2931,10 +3077,12 @@ class AKSPreviewManagedClusterCreateDecorator(AKSManagedClusterCreateDecorator):
         mc = self.set_up_node_restriction(mc)
         # set up image cleaner
         mc = self.set_up_image_cleaner(mc)
+        # set up image integrity
+        mc = self.set_up_image_integrity(mc)
         # set up cluster snapshot
         mc = self.set_up_creationdata_of_cluster_snapshot(mc)
-        # set up ingress web app routing profile
-        mc = self.set_up_ingress_web_app_routing(mc)
+        # set up app routing profile
+        mc = self.set_up_app_routing_profile(mc)
         # set up workload auto scaler profile
         mc = self.set_up_workload_auto_scaler_profile(mc)
         # set up vpa
@@ -2959,6 +3107,8 @@ class AKSPreviewManagedClusterCreateDecorator(AKSManagedClusterCreateDecorator):
         mc = self.set_up_metrics_profile(mc)
         # set up for azure container storage
         mc = self.set_up_azure_container_storage(mc)
+        # set up node provisioning profile
+        mc = self.set_up_node_provisioning_profile(mc)
 
         # DO NOT MOVE: keep this at the bottom, restore defaults
         mc = self._restore_defaults_in_mc(mc)
@@ -3266,6 +3416,10 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
         :return: the ManagedCluster object
         """
         self._ensure_mc(mc)
+
+        network_plugin = self.context._get_network_plugin()
+        if network_plugin:
+            mc.network_profile.network_plugin = network_plugin
 
         network_plugin_mode = self.context.get_network_plugin_mode()
         if network_plugin_mode:
@@ -3591,11 +3745,20 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
         """
         self._ensure_mc(mc)
 
+        enable_image_integrity = self.context.get_enable_image_integrity()
         disable_image_integrity = self.context.get_disable_image_integrity()
 
         # no image integrity related changes
-        if not disable_image_integrity:
+        if not enable_image_integrity and not disable_image_integrity:
             return mc
+        if enable_image_integrity and disable_image_integrity:
+            raise MutuallyExclusiveArgumentError(
+                "Cannot specify --enable-image-integrity and "
+                "--disable-image-integrity at the same time."
+            )
+        shouldEnable_image_integrity = False
+        if enable_image_integrity:
+            shouldEnable_image_integrity = True
 
         if mc.security_profile is None:
             mc.security_profile = self.models.ManagedClusterSecurityProfile()
@@ -3606,7 +3769,7 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
             image_integrity_profile = self.models.ManagedClusterSecurityProfileImageIntegrity()
             mc.security_profile.image_integrity = image_integrity_profile
 
-        image_integrity_profile.enabled = False
+        image_integrity_profile.enabled = shouldEnable_image_integrity
 
         return mc
 
@@ -3965,6 +4128,19 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
 
         return mc
 
+    def update_node_provisioning_mode(self, mc: ManagedCluster) -> ManagedCluster:
+        self._ensure_mc(mc)
+
+        mode = self.context.get_node_provisioning_mode()
+        if mode is not None:
+            if mc.node_provisioning_profile is None:
+                mc.node_provisioning_profile = self.models.ManagedClusterNodeProvisioningProfile()
+
+            # set mode
+            mc.node_provisioning_profile.mode = mode
+
+        return mc
+
     def update_metrics_profile(self, mc: ManagedCluster) -> ManagedCluster:
         """Updates the metricsProfile field of the managed cluster
 
@@ -3973,6 +4149,91 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
         self._ensure_mc(mc)
 
         mc = self.update_cost_analysis(mc)
+
+        return mc
+
+    def update_app_routing_profile(self, mc: ManagedCluster) -> ManagedCluster:
+        """Update app routing profile for the ManagedCluster object.
+
+        :return: the ManagedCluster object
+        """
+        from azure.cli.command_modules.keyvault.custom import set_policy
+        from azext_aks_preview._client_factory import get_keyvault_client
+
+        self._ensure_mc(mc)
+
+        # get parameters from context
+        enable_app_routing = self.context.get_enable_app_routing()
+        enable_keyvault_secret_provider = self.context.get_enable_kv()
+        attach_zones = self.context.get_attach_zones()
+        dns_zone_resource_ids = self.context.get_dns_zone_resource_ids_from_input()
+        add_dns_zone = self.context.get_add_dns_zone()
+        delete_dns_zone = self.context.get_delete_dns_zone()
+        update_dns_zone = self.context.get_update_dns_zone()
+
+        # update ManagedCluster object with app routing settings
+        mc.ingress_profile = mc.ingress_profile or self.models.ManagedClusterIngressProfile()
+        mc.ingress_profile.web_app_routing = mc.ingress_profile.web_app_routing or self.models.ManagedClusterIngressProfileWebAppRouting()
+        if enable_app_routing is not None:
+                if mc.ingress_profile.web_app_routing.enabled == enable_app_routing:
+                    error_message = 'App Routing is already enabled.\n' if enable_app_routing else 'App Routing is already disabled.\n'
+                    raise CLIError(error_message)
+                mc.ingress_profile.web_app_routing.enabled = enable_app_routing
+        # update ManagedCluster object with keyvault-secret-provider settings
+        if enable_keyvault_secret_provider:
+            mc.addon_profiles = mc.addon_profiles or {}
+            if not mc.addon_profiles.get(CONST_AZURE_KEYVAULT_SECRETS_PROVIDER_ADDON_NAME):
+                mc.addon_profiles[CONST_AZURE_KEYVAULT_SECRETS_PROVIDER_ADDON_NAME] = self.models.ManagedClusterAddonProfile(
+                    enabled=True, config={CONST_SECRET_ROTATION_ENABLED: "false", CONST_ROTATION_POLL_INTERVAL: "2m"})
+            elif not mc.addon_profiles[CONST_AZURE_KEYVAULT_SECRETS_PROVIDER_ADDON_NAME].enabled:
+                mc.addon_profiles[CONST_AZURE_KEYVAULT_SECRETS_PROVIDER_ADDON_NAME].enabled = True
+
+        # modify DNS zone resource IDs
+        if dns_zone_resource_ids:
+            if mc.ingress_profile and mc.ingress_profile.web_app_routing and mc.ingress_profile.web_app_routing.enabled:
+                if add_dns_zone:
+                    if mc.ingress_profile.web_app_routing.dns_zone_resource_ids is None:
+                        mc.ingress_profile.web_app_routing.dns_zone_resource_ids = []
+                    mc.ingress_profile.web_app_routing.dns_zone_resource_ids.extend(dns_zone_resource_ids)
+                    if attach_zones:
+                        try:
+                            for dns_zone in dns_zone_resource_ids:
+                                if not add_role_assignment(self.cmd, 'DNS Zone Contributor', mc.ingress_profile.web_app_routing.identity.object_id, False, scope=dns_zone):
+                                    logger.warning(
+                                        'Could not create a role assignment for App Routing. '
+                                        'Are you an Owner on this subscription?')
+                        except Exception as ex:
+                            raise CLIError(f'Error in granting dns zone permisions to managed identity: {ex}\n')
+                elif delete_dns_zone:
+                    if mc.ingress_profile.web_app_routing.dns_zone_resource_ids:
+                        dns_zone_resource_ids = [x for x in mc.ingress_profile.web_app_routing.dns_zone_resource_ids if x not in dns_zone_resource_ids]
+                        mc.ingress_profile.web_app_routing.dns_zone_resource_ids = dns_zone_resource_ids
+                    else:
+                        raise CLIError('No DNS zone is used by App Routing.\n')
+                elif update_dns_zone:
+                    mc.ingress_profile.web_app_routing.dns_zone_resource_ids = dns_zone_resource_ids
+                    if attach_zones:
+                        try:
+                            for dns_zone in dns_zone_resource_ids:
+                                if not add_role_assignment(self.cmd, 'DNS Zone Contributor', mc.ingress_profile.web_app_routing.identity.object_id, False, scope=dns_zone):
+                                    logger.warning(
+                                        'Could not create a role assignment for App Routing. '
+                                        'Are you an Owner on this subscription?')
+                        except Exception as ex:
+                            raise CLIError(f'Error in granting dns zone permisions to managed identity: {ex}\n')
+            else:
+                raise CLIError('App Routing must be enabled to modify DNS zone resource IDs.\n')
+
+        return mc
+
+    def update_node_provisioning_profile(self, mc: ManagedCluster) -> ManagedCluster:
+        """Updates the nodeProvisioningProfile field of the managed cluster
+
+        :return: the ManagedCluster object
+        """
+        self._ensure_mc(mc)
+
+        mc = self.update_node_provisioning_mode(mc)
 
         return mc
 
@@ -4037,6 +4298,8 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
         mc = self.update_metrics_profile(mc)
         # update azure container storage
         mc = self.update_azure_container_storage(mc)
+        # update node provisioning profile
+        mc = self.update_node_provisioning_profile(mc)
 
         return mc