+++ src/devcenter/azext_devcenter/aaz/latest/devcenter/admin/devbox_definition/_update.py
@@ -22,9 +22,9 @@ class Update(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2023-06-01-preview",
+        "version": "2023-10-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/devcenters/{}/devboxdefinitions/{}", "2023-06-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/devcenters/{}/devboxdefinitions/{}", "2023-10-01-preview"],
         ]
     }
 
@@ -52,12 +52,22 @@ class Update(AAZCommand):
             help="The name of the dev box definition.",
             required=True,
             id_part="child_name_1",
+            fmt=AAZStrArgFormat(
+                pattern="^[a-zA-Z0-9][a-zA-Z0-9-_.]{2,62}$",
+                max_length=63,
+                min_length=3,
+            ),
         )
         _args_schema.dev_center_name = AAZStrArg(
             options=["-d", "--dev-center", "--dev-center-name"],
             help="The name of the dev center. Use `az configure -d dev-center=<dev_center_name>` to configure a default.",
             required=True,
             id_part="name",
+            fmt=AAZStrArgFormat(
+                pattern="^[a-zA-Z0-9][a-zA-Z0-9-]{2,25}$",
+                max_length=26,
+                min_length=3,
+            ),
         )
         _args_schema.resource_group = AAZResourceGroupNameArg(
             required=True,
@@ -223,7 +233,7 @@ class Update(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2023-06-01-preview",
+                    "api-version", "2023-10-01-preview",
                     required=True,
                 ),
             }
@@ -326,7 +336,7 @@ class Update(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2023-06-01-preview",
+                    "api-version", "2023-10-01-preview",
                     required=True,
                 ),
             }
@@ -491,10 +501,6 @@ class _UpdateHelper:
         properties.sku = AAZObjectType(
             flags={"required": True},
         )
-        properties.validation_error_details = AAZListType(
-            serialized_name="validationErrorDetails",
-            flags={"read_only": True},
-        )
         properties.validation_status = AAZStrType(
             serialized_name="validationStatus",
         )
@@ -511,13 +517,6 @@ class _UpdateHelper:
         )
         sku.size = AAZStrType()
         sku.tier = AAZStrType()
-
-        validation_error_details = _schema_dev_box_definition_read.properties.validation_error_details
-        validation_error_details.Element = AAZObjectType()
-
-        _element = _schema_dev_box_definition_read.properties.validation_error_details.Element
-        _element.code = AAZStrType()
-        _element.message = AAZStrType()
 
         system_data = _schema_dev_box_definition_read.system_data
         system_data.created_at = AAZStrType(